Add logging, update requirement files, create EDA notebook
This commit includes three major changes.
First, logging-related code is added to the configuration for better tracking and issue debugging. This was done by importing the 'logging' and 'rich' libraries and setting up a logger.
Second, four new requirements (pandas, xgboost, polars, rich) have been added to the requirements.txt file. This ensures that all necessary dependencies will be installed for any new developers or environments for project consistency and compatibility.
Finally, a new Jupyter notebook (eda.ipynb) is created under the 'research' folder. This contains an exploratory data analysis (EDA) on the dataset, which should serve as a useful point of reference when modeling and evaluating predictions.